IVSCC auto recon 1003

+++ hackathon/MK/v3d_imgManagerMK/imgAnalyzer/ImgAnalyzer.h
@@ -48,9 +48,9 @@ inline void ImgAnalyzer::ChebyshevCenter(set<vector<int>> allCoords, float cente
 		float currLengthSum = 0;
 		for (set<vector<int>>::iterator checkCoordIt = allCoords.begin(); checkCoordIt != allCoords.end(); ++checkCoordIt)
 		{
-			float length = sqrt((checkCoordIt->at(0) - allCoordIt->at(0)) * (checkCoordIt->at(0) - allCoordIt->at(0)) +
-				(checkCoordIt->at(1) - allCoordIt->at(1)) * (checkCoordIt->at(1) - allCoordIt->at(1)) +
-				(checkCoordIt->at(2) - allCoordIt->at(2)) * (checkCoordIt->at(2) - allCoordIt->at(2)));
+			float length = sqrt(float((checkCoordIt->at(0) - allCoordIt->at(0)) * (checkCoordIt->at(0) - allCoordIt->at(0)) +
+									  (checkCoordIt->at(1) - allCoordIt->at(1)) * (checkCoordIt->at(1) - allCoordIt->at(1)) +
+									  (checkCoordIt->at(2) - allCoordIt->at(2)) * (checkCoordIt->at(2) - allCoordIt->at(2))));
 
 			currLengthSum += length;
 		}
